added controllers and models
