+++ models/advertisingModels.js
@@ -1,0 +1,5 @@
+const registeredAddresses = [
+  "157 Nguyễn Đình Chính, Phường 11, Quận Phú Nhuận",
+];
+
+export { registeredAddresses };
