+++ models/user.js
@@ -1,5 +1,7 @@
 import mongoose from "mongoose";
-mongoose.connect("mongodb+srv://nhom09:atlas123@cluster0.hntnfkf.mongodb.net/");
+await mongoose.connect(
+  "mongodb+srv://nhom09:atlas123@cluster0.hntnfkf.mongodb.net/"
+);
 
 const userSchema = new mongoose.Schema({
   email: { type: String, unique: true, required: true },
